requirements.txt
torch
torchvision
numpy
pandas
Pillow
pyyaml
